Added DraftService prototype. Other changes
