+++ backend/Models/API/Common/IObjectWithIndex.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace backend.Models.API.Common
+{
+	public interface IObjectWithIndex
+	{
+		[Key]
+		public int Id { get; set; }
+	}
+}
