+++ backend/Models/API/Post/DeletePostRequest.cs
@@ -1,8 +1,9 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using backend.Models.API.Common;
+using System.ComponentModel.DataAnnotations;
 
 namespace backend.Models.API.Requests
 {
-	public class DeletePostRequest
+	public class DeletePostRequest 
 	{
 		[Required]
 		public int Id { get; set; }
