+++ backend/Models/API/User/UserPatchRequest.cs
@@ -1,6 +1,8 @@
-﻿namespace backend.Models.API.User
+﻿using backend.Models.API.Common;
+
+namespace backend.Models.API.User
 {
-	public class UserPatchRequest
+	public class UserPatchRequest //:
 	{
 		public int? Id { get; set; }
 		public string? NewName { get; set; }
