+++ backend/appsettings.json
@@ -40,5 +40,12 @@
   },
   "PasswordsCryptographyServiceSettings": {
     "saltSizeBytes": 128
+  },
+  "DraftServiceSettings": {
+    "maxEntitiesStoredInRam": 100,
+    "maxTimeEntityStoredInRamMinutes": 10,
+    "maxEntitiesStoredInDb": 10000,
+    "maxTimeEntityStoredInDbMinutes": 43200, // 30 days
+    "updateStoredEntitiesIntervalMinutes": 10
   }
 }
